requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP requests for API calls
httpx==0.25.1
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Templates
jinja2==3.1.2

# Utilities
python-magic-bin==0.4.14; platform_system == "Windows"
python-magic==0.4.27; platform_system != "Windows"

